# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Shared memory layer for security operations: episodic observations, versioned facts, hybrid recall, graph traversal.",
        "title": "SOC Memory Layer API",
        "contact": {
            "name": "SOC Platform Team"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/enrich": {
            "get": {
                "security": [
                    {
                        "IdentityHeaders": []
                    }
                ],
                "description": "Entity card for one key: profile, facts, observations, and graph neighbors.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recall"
                ],
                "summary": "Enrich an entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entity type (e.g. ioc_ip, ioc_domain)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Entity lookup key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entity card; found=false with empty lists when unknown",
                        "schema": {
                            "$ref": "#/definitions/internal_api.enrichResp"
                        }
                    },
                    "400": {
                        "description": "Missing type/key or storage validation failure",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "502": {
                        "description": "Storage temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    }
                }
            }
        },
        "/v1/facts": {
            "post": {
                "security": [
                    {
                        "IdentityHeaders": []
                    }
                ],
                "description": "Assert a subject-predicate-object fact; agents propose, humans write active versions directly.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "write"
                ],
                "summary": "Assert a fact",
                "parameters": [
                    {
                        "description": "Fact payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.assertFactReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Written fact (proposed or active per actor type)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.factResp"
                        }
                    },
                    "400": {
                        "description": "Invalid body or oversized predicate/value",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "429": {
                        "description": "Agent write budget exhausted (Retry-After set)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "502": {
                        "description": "Storage temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    }
                }
            }
        },
        "/v1/facts/{id}/promote": {
            "post": {
                "security": [
                    {
                        "IdentityHeaders": []
                    }
                ],
                "description": "Transition a proposed fact to active; human-gated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "write"
                ],
                "summary": "Promote a fact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fact ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Promoted fact with the current version ID",
                        "schema": {
                            "$ref": "#/definitions/internal_api.transitionResp"
                        }
                    },
                    "400": {
                        "description": "Invalid fact ID",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "403": {
                        "description": "Only human actors may promote (human_gated)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "404": {
                        "description": "Fact not found or ID consumed by an earlier transition",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "409": {
                        "description": "Fact is not in proposed status",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "429": {
                        "description": "Agent write budget exhausted (Retry-After set)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "502": {
                        "description": "Storage temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    }
                }
            }
        },
        "/v1/facts/{id}/retract": {
            "post": {
                "security": [
                    {
                        "IdentityHeaders": []
                    }
                ],
                "description": "Mark a fact retracted with an optional reason; human-gated. Empty request body allowed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "write"
                ],
                "summary": "Retract a fact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fact ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Optional reason",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.retractReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retracted fact with the current version ID",
                        "schema": {
                            "$ref": "#/definitions/internal_api.transitionResp"
                        }
                    },
                    "400": {
                        "description": "Invalid fact ID or malformed body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "403": {
                        "description": "Only human actors may retract (human_gated)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "404": {
                        "description": "Fact not found or ID consumed by an earlier transition",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "409": {
                        "description": "Fact already retracted",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "429": {
                        "description": "Agent write budget exhausted (Retry-After set)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "502": {
                        "description": "Storage temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    }
                }
            }
        },
        "/v1/observations": {
            "post": {
                "security": [
                    {
                        "IdentityHeaders": []
                    }
                ],
                "description": "Store a new episodic observation; actor identity comes from X-Actor-* headers, never the body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "write"
                ],
                "summary": "Record an observation",
                "parameters": [
                    {
                        "description": "Observation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.createObservationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created observation",
                        "schema": {
                            "$ref": "#/definitions/internal_api.observationResp"
                        }
                    },
                    "400": {
                        "description": "Invalid body, oversized content, or invalid ts",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "429": {
                        "description": "Agent write budget exhausted (Retry-After set)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "502": {
                        "description": "Storage temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    }
                }
            }
        },
        "/v1/similar": {
            "get": {
                "security": [
                    {
                        "IdentityHeaders": []
                    }
                ],
                "description": "Semantic + keyword hybrid search over the scope's observations, ranked.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recall"
                ],
                "summary": "Hybrid recall search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Natural-language query text",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Maximum hits in [1,50] (default 10)",
                        "name": "k",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ranked hits (possibly empty)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.searchHitJSON"
                            }
                        }
                    },
                    "400": {
                        "description": "Missing q or out-of-range/non-integer k",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "502": {
                        "description": "Storage temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    }
                }
            }
        },
        "/v1/timeline": {
            "get": {
                "security": [
                    {
                        "IdentityHeaders": []
                    }
                ],
                "description": "Chronological events scoped to exactly one of case_id or entity_id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recall"
                ],
                "summary": "Case or entity timeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Case UUID (exactly one of case_id/entity_id)",
                        "name": "case_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Entity ID (exactly one of case_id/entity_id)",
                        "name": "entity_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset into the result set (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Events, newest first",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.eventJSON"
                            }
                        }
                    },
                    "400": {
                        "description": "Both/neither selector given or non-integer paging",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "502": {
                        "description": "Storage temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    }
                }
            }
        },
        "/v1/traverse": {
            "get": {
                "security": [
                    {
                        "IdentityHeaders": []
                    }
                ],
                "description": "Walk relations outward from an entity up to hops (1-3); degrades to at most 1 hop without Dgraph.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recall"
                ],
                "summary": "Graph traversal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start entity lookup key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start entity type (e.g. ioc_domain)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Restrict edges to one relation type",
                        "name": "relation",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Hops in [1,3] (default 1)",
                        "name": "hops",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Walked paths from the start entity",
                        "schema": {
                            "$ref": "#/definitions/internal_api.traverseResp"
                        }
                    },
                    "400": {
                        "description": "Missing type/key, bad hops, or malformed relation",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    },
                    "502": {
                        "description": "Storage temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_api.assertFactReq": {
            "type": "object",
            "properties": {
                "client_event_id": {
                    "type": "string"
                },
                "confidence": {
                    "type": "number"
                },
                "object_id": {
                    "type": "string"
                },
                "object_value": {
                    "type": "string"
                },
                "predicate": {
                    "type": "string"
                },
                "source_obs": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                }
            }
        },
        "internal_api.createObservationReq": {
            "type": "object",
            "properties": {
                "case_id": {
                    "type": "string"
                },
                "client_event_id": {
                    "type": "string"
                },
                "confidentiality": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "ts": {
                    "description": "RFC3339; omitted = now",
                    "type": "string"
                }
            }
        },
        "internal_api.enrichResp": {
            "type": "object",
            "properties": {
                "entity": {
                    "$ref": "#/definitions/internal_api.entityJSON"
                },
                "facts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.factViewJSON"
                    }
                },
                "found": {
                    "type": "boolean"
                },
                "neighbors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.neighborJSON"
                    }
                },
                "observations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.obsViewJSON"
                    }
                }
            }
        },
        "internal_api.entityJSON": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "first_seen": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "last_seen": {
                    "type": "string"
                },
                "scope": {
                    "description": "true origin scope of the hydrated row",
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "internal_api.errBody": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_api.errInner"
                }
            }
        },
        "internal_api.errInner": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_api.eventJSON": {
            "type": "object",
            "properties": {
                "actor_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "origin_scope": {
                    "description": "scope the underlying row lives in",
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "ts": {
                    "type": "string"
                }
            }
        },
        "internal_api.factResp": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "object_value": {
                    "type": "string"
                },
                "predicate": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "valid_from": {
                    "type": "string"
                },
                "written_by": {
                    "type": "string"
                }
            }
        },
        "internal_api.factViewJSON": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "object_value": {
                    "type": "string"
                },
                "origin_scope": {
                    "description": "scope the fact was asserted in",
                    "type": "string"
                },
                "predicate": {
                    "type": "string"
                },
                "source_obs": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "valid_from": {
                    "type": "string"
                },
                "written_by": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.neighborJSON": {
            "type": "object",
            "properties": {
                "direction": {
                    "type": "string"
                },
                "entity_id": {
                    "type": "string"
                },
                "relation": {
                    "type": "string"
                }
            }
        },
        "internal_api.obsViewJSON": {
            "type": "object",
            "properties": {
                "excerpt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "origin_scope": {
                    "description": "scope the observation was recorded in",
                    "type": "string"
                },
                "ts": {
                    "type": "string"
                }
            }
        },
        "internal_api.observationResp": {
            "type": "object",
            "properties": {
                "actor_id": {
                    "type": "string"
                },
                "actor_type": {
                    "type": "string"
                },
                "embedded": {
                    "type": "boolean"
                },
                "entity_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "ts": {
                    "type": "string"
                }
            }
        },
        "internal_api.pathJSON": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.entityJSON"
                    }
                },
                "relations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "internal_api.retractReq": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
        "internal_api.searchHitJSON": {
            "type": "object",
            "properties": {
                "excerpt": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "matched_by": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "obs_id": {
                    "type": "string"
                },
                "scope": {
                    "description": "originating scope (org-wide recall attribution)",
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "ts": {
                    "type": "string"
                }
            }
        },
        "internal_api.transitionResp": {
            "type": "object",
            "properties": {
                "fact": {
                    "$ref": "#/definitions/internal_api.factResp"
                },
                "transitioned_by": {
                    "type": "string"
                }
            }
        },
        "internal_api.traverseResp": {
            "type": "object",
            "properties": {
                "paths": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.pathJSON"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "IdentityHeaders": {
            "description": "Plus X-Actor-ID and X-Scope headers; identity never travels in bodies.",
            "type": "apiKey",
            "name": "X-Actor-Type",
            "in": "header"
        }
    }
}